DI, Implementação de Notificador de Erros e Melhoria na Service

+++ Back/src/EntregaSegura.Application/Services/CondominioService.cs
@@ -1,4 +1,5 @@
 using EntregaSegura.Application.Interfaces;
+using EntregaSegura.Application.Notifications;
 using EntregaSegura.Domain.Entities;
 using EntregaSegura.Domain.Interfaces.Repositories;
 using EntregaSegura.Domain.Validators;
@@ -10,7 +11,9 @@ public class CondominioService : BaseService, ICondominioService
 {
     private readonly ICondominioRepository _condominioRepository;
 
-    public CondominioService(IUnitOfWork unitOfWork, ICondominioRepository condominioRepository) : base(unitOfWork)
+    public CondominioService(ICondominioRepository condominioRepository,
+                             IUnitOfWork unitOfWork,
+                             NotificadorErros notificadorErros) : base(unitOfWork, notificadorErros)
     {
         _condominioRepository = condominioRepository;
     }
@@ -18,6 +21,18 @@ public class CondominioService : BaseService, ICondominioService
     public async Task Adicionar(Condominio condominio)
     {
         if(!ExecutarValidacao(new CondominioValidator(), condominio)) return;
+
+        if(_condominioRepository.BuscarAsync(c => c.CNPJ == condominio.CNPJ).Result.Any())
+        {
+            Notificar("Já existe um condomínio com este CNPJ.");
+            return;
+        }
+
+        if(_condominioRepository.BuscarAsync(c => c.Nome == condominio.Nome).Result.Any())
+        {
+            Notificar("Já existe um condomínio com este nome.");
+            return;
+        }
 
         _condominioRepository.Adicionar(condominio);
         await CommitAsync();
@@ -27,13 +42,29 @@ public class CondominioService : BaseService, ICondominioService
     {
         if(!ExecutarValidacao(new CondominioValidator(), condominio)) return;
 
+        if(_condominioRepository.BuscarAsync(c => c.CNPJ == condominio.CNPJ && c.Id != condominio.Id).Result.Any())
+        {
+            Notificar("Já existe um condomínio com este CNPJ.");
+            return;
+        }
+
         _condominioRepository.Atualizar(condominio);
         await CommitAsync();
     }
 
     public async Task Remover(Guid id)
     {
+        if(_condominioRepository.ObterCondominioComUnidadesAsync(id).Result.Unidades.Any())
+        {
+            Notificar("O condomínio possui unidades cadastradas!");
+            return;
+        }
         _condominioRepository.Remover(id);
         await CommitAsync();
+    }
+
+    public void Dispose()
+    {
+        _unitOfWork?.Dispose();
     }
 }
